requirements.txt
certifi==2020.12.5
chardet==4.0.0
filelock==3.0.12
gdown==3.12.2
idna==2.10
numpy==1.20.0
opencv-python==4.5.1.48
PySocks==1.7.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
tqdm==4.56.0
urllib3==1.26.3
